+++ src/Core/Dfe.Complete.Domain/Enums/NoteTaskIdentifier.cs
@@ -23,5 +23,17 @@ public enum NoteTaskIdentifier
 
     [Description("supplemental_funding_agreement")]
     [DisplayDescription("Supplemental funding agreement")]
-    SupplementalFundingAgreement = 5
+    SupplementalFundingAgreement = 5,
+
+    [Description("article_of_association")]
+    [DisplayDescription("Articles of association")]
+    ArticleOfAssociation = 6,
+
+    [Description("deed_of_novation_and_variation")]
+    [DisplayDescription("Deed of novation and variation\r\n")]
+    DeedOfNovationAndVariation = 7,
+
+    [Description("deed_of_variation")]
+    [DisplayDescription("Deed of variation")]
+    DeedOfVariation = 8,
 }