+++ src/Frontend/Dfe.Complete/Security/CustomPolicies.cs
@@ -18,17 +18,6 @@ public static class CustomPolicies
                     (user.IsInRole(UserRolesConstants.RegionalCaseworkServices) && !user.IsInRole(UserRolesConstants.ManageTeam));
             });
         },
-        [UserPolicyConstants.CanViewYourProjects] = builder =>
-        {
-            builder.RequireAuthenticatedUser();
-            builder.RequireAssertion(context =>
-            {
-                var user = context.User;
-                return
-                    user.IsInRole(UserRolesConstants.RegionalDeliveryOfficer) ||
-                    (user.IsInRole(UserRolesConstants.RegionalCaseworkServices) && !user.IsInRole(UserRolesConstants.ManageTeam));
-            });
-        },
         [UserPolicyConstants.CanViewTeamProjectsUnassigned] = builder =>
         {
             builder.RequireAuthenticatedUser();