+++ CHANGELOG.md
@@ -16,11 +16,13 @@ Security in case of vulnerabilities.
 
 ## Unreleased
 
-### Added
-
 ### Changed
+- Complete project button will hide when users don't have access
+- Sort "Your projects in progress" in ascending date order
 
 ### Fixed
+- Complete project post no longer throws antiforgery error
+
 
 See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-09-03.710...main) for everything awaiting release